refactor/feat: Декомпозиция HeaderNav. Ссылки sing in/sign out отображаются в зависимости от статуса авторизации
