+++ src/components/header-nav/components/avatar/avatar.tsx
@@ -1,0 +1,7 @@
+function Avatar(): JSX.Element {
+  return (
+    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+  );
+}
+
+export default Avatar;
