+++ src/components/header-nav/components/user/user.tsx
@@ -1,0 +1,22 @@
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../../../const';
+import Avatar from '../avatar/avatar';
+
+function User():JSX.Element{
+  return (
+    <li className="header__nav-item user">
+      <Link
+        className="header__nav-link header__nav-link--profile"
+        to={AppRoute.Favorites}
+      >
+        <Avatar/>
+        <span className="header__user-name user__name">
+                  Oliver.conner@gmail.com
+        </span>
+        <span className="header__favorite-count">3</span>
+      </Link>
+    </li>
+  );
+}
+
+export default User;
